Refactor files structure + split test files
